services/ingest-hltv/pyproject.toml
[tool.poetry]
name = "ingest-hltv"
version = "0.1.0"
description = "HLTV ingest and realtime scraper"
authors = ["faztyG <verselabproject@gmail.com>"]
readme = "README.md"
packages = [{ include = "ingest_hltv" }]

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
httpx = "^0.28.1"
orjson = "^3.11.2"
nats-py = "^2.11.0"
redis = "^6.4.0"
playwright = "^1.47.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
